backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
anthropic>=0.28.0
apscheduler>=3.10.4
httpx>=0.27.0
python-dotenv>=1.0.1
aiosqlite>=0.20.0
